requirements.txt
djangorestframework
numpy
requests
beautifulsoup4